feat: Move AI provider model fetching to server-side for Docker network compatibility
## Problem
Previously, the `fetchProviderModels` function made HTTP requests directly from the browser to fetch AI provider model lists. This meant Docker internal hostnames (like `http://ollama:11434`) were not accessible from the browser.

## Solution
- Move model fetching logic from frontend to backend tRPC endpoint
- Frontend now calls `api.ai.fetchProviderModels.mutate()` which triggers server-side fetch
- Server-side fetch can resolve Docker internal hostnames
- Supports existing HTTP proxy configuration

## Changes
- `server/routerTrpc/ai.ts`: Add `fetchProviderModels` tRPC endpoint
- `shared/lib/modelTemplates.ts`: New shared module for model templates and inferModelCapabilities
- `app/src/store/aiSettingStore.tsx`: Simplify to call backend API
- `app/src/components/BlinkoSettings/AiSetting/constants.tsx`: Re-export from shared module
- `test-docker/`: Add Docker test environment with mock OpenAI API

## Testing
See `test-docker/README.md` for Docker internal network testing instructions.
